refactor + wip rooms
