--- model.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-
-class LoggedInUser(BaseModel):
-    username: str
-    uuid: str
-
-    room: str | None = None
-
-
