+++ src/models/model.py
@@ -1,0 +1,12 @@
+from pydantic import BaseModel
+
+
+class LoggedInUser(BaseModel):
+    username: str
+    uuid: str
+    room_code: str | None = None
+
+
+class MojangInfo(BaseModel):
+    serverID: str
+    username: str
